Ignore inexistent results and filter by config.
The UI used to push inexistent results, which never happened in practice
but now does with the use of multiple configs (Bigloo has a few missing
results). The UI now allows selecting which config to view results for.
The app is also configurable by top-level variables which allow
customizing between systems ('gambit' and 'zipi' for example).